@@ -4,51 +4,48 @@
   "homepage": "/",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.12.0",
-    "@fullcalendar/core": "^4.3.1",
-    "@fullcalendar/daygrid": "^4.3.0",
-    "@fullcalendar/react": "^4.3.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@tinymce/tinymce-react": "^3.5.0",
-    "bootstrap": "^4.4.1",
-    "bs-custom-file-input": "^1.3.4",
-    "chart.js": "^2.9.3",
-    "firebase": "^10.13.0",
-    "flag-icon-css": "^3.4.5",
-    "gh-pages": "^6.1.1",
-    "ionicons": "^4.6.3",
-    "nouislider-react": "^3.3.7",
-    "react": "^16.12.0",
-    "react-bootstrap": "^1.0.0-beta.16",
-    "react-bootstrap-table-next": "^3.3.5",
-    "react-bootstrap-table2-paginator": "^2.1.0",
-    "react-bootstrap-table2-toolkit": "^2.1.2",
-    "react-chartjs-2": "^2.9.0",
-    "react-color": "^2.18.0",
-    "react-datepicker": "^2.13.0",
-    "react-dom": "^16.12.0",
-    "react-firebase-hooks": "^5.1.1",
-    "react-input-mask": "^2.0.4",
-    "react-multistep": "^3.7.3",
-    "react-perfect-scrollbar": "^1.5.3",
-    "react-quill": "^1.3.3",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "react-select": "^3.0.8",
-    "react-simplemde-editor": "^4.1.0",
-    "sass": "^1.77.8",
-    "typicons.font": "^2.0.9"
+    "@fortawesome/fontawesome-free": "^6.6.0",
+    "@fullcalendar/core": "^6.1.15",
+    "@fullcalendar/daygrid": "^6.1.15",
+    "@fullcalendar/react": "^6.1.8",
+    "@testing-library/jest-dom": "^6.3.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@tinymce/tinymce-react": "^4.5.0",
+    "bootstrap": "^5.3.1",
+    "bs-custom-file-input": "^1.4.0",
+    "chart.js": "^4.4.0",
+    "firebase": "^10.4.1",
+    "flag-icon-css": "^4.2.0",
+    "gh-pages": "^6.3.1",
+    "ionicons": "^7.4.0",
+    "nouislider-react": "^4.0.0",
+    "react": "^18.3.1",
+    "react-bootstrap": "^5.0.0-beta.0",
+    "react-bootstrap-table-next": "^4.0.3",
+    "react-bootstrap-table2-paginator": "^2.2.3",
+    "react-bootstrap-table2-toolkit": "^3.0.0",
+    "react-chartjs-2": "^5.2.0",
+    "react-color": "^2.19.3",
+    "react-datepicker": "^4.17.1",
+    "react-dom": "^18.3.1",
+    "react-firebase-hooks": "^6.1.1",
+    "react-input-mask": "^3.1.1",
+    "react-multistep": "^4.0.0",
+    "react-perfect-scrollbar": "^1.5.9",
+    "react-quill": "^2.0.0-beta.2",
+    "react-router-dom": "^6.16.0",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.8.1",
+    "react-simplemde-editor": "^5.0.0",
+    "sass": "^1.77",
+    "typicons.font": "^2.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "export NODE_OPTIONS=--openssl-legacy-provider && react-scripts build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "engines": {
-    "node": "16.x"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -66,6 +63,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@types/react-router-dom": "^5.3.3",
     "compass-mixins": "^0.12.10"
   }
