@@ -1,55 +1,41 @@
-import React, { Component } from 'react';
-import { withRouter } from 'react-router-dom';
+import React, { useEffect, useMemo } from 'react';
+import { useLocation } from 'react-router-dom';
 import './App.scss';
 import AppRoutes from './AppRoutes';
 import Header from './shared/Header';
 import Footer from './shared/Footer';
 
-class App extends Component {
-  state = {}
-  componentDidMount() {
-    this.onRouteChanged();
-  }
-  render () {
-    let headerComponent = !this.state.isFullPageLayout ? <Header/> : '';
-    let footerComponent = !this.state.isFullPageLayout ? <Footer/> : '';
-    return (
-      <div>
-        { headerComponent }
-        <div className="az-content-wrapper">
-          <AppRoutes/>
-        </div>
-        { footerComponent }
-      </div>
-    );
-  }
+const App = () => {
+  const location = useLocation();
 
-  componentDidUpdate(prevProps) {
-    if (this.props.location !== prevProps.location) {
-      this.onRouteChanged();
-    }
-  }
-
-  onRouteChanged() {
-    console.log("ROUTE CHANGED");
-    window.scrollTo(0, 0);
+  const isFullPageLayout = useMemo(() => {
     const fullPageLayoutRoutes = ['/general-pages/signin', '/general-pages/signup', '/general-pages/page-404'];
-    for ( let i = 0; i < fullPageLayoutRoutes.length; i++ ) {
-      if (this.props.location.pathname === fullPageLayoutRoutes[i]) {
-        this.setState({
-          isFullPageLayout: true
-        })
-        document.querySelector('.az-content-wrapper').classList.add('p-0');
-        break;
+    return fullPageLayoutRoutes.includes(location.pathname);
+  }, [location.pathname]);
+
+  useEffect(() => {
+    // Scroll to top on route change
+    window.scrollTo(0, 0);
+
+    const contentWrapper = document.querySelector('.az-content-wrapper');
+    if (contentWrapper) {
+      if (isFullPageLayout) {
+        contentWrapper.classList.add('p-0');
       } else {
-        this.setState({
-          isFullPageLayout: false
-        })
-        document.querySelector('.az-content-wrapper').classList.remove('p-0');
+        contentWrapper.classList.remove('p-0');
       }
     }
-  }
+  }, [isFullPageLayout]);
 
-}
+  return (
+    <div>
+      {!isFullPageLayout && <Header />}
+      <div className="az-content-wrapper">
+        <AppRoutes />
+      </div>
+      {!isFullPageLayout && <Footer />}
+    </div>
+  );
+};
 
-export default withRouter(App);
+export default App;
