@@ -2,33 +2,34 @@ import React, { useState, useEffect } from 'react';
 import TabContents from './TabContents';
 import ExecutiveView from './ExecutiveView';
 import RiskView from './RiskView';
-  const staticData = {
-    overview: {
-      title: 'Overview Data',
-      content: 'This is some static executive view data. Replace with actual content.'
-    },
-    executive: {
-      title: 'Executive View Data',
-      content: 'This is some static executive view data. Replace with actual content.'
-    },
-    risk: {
-      title: 'Risk & Vulnerability View Data',
-      content: 'This is some static risk & vulnerability view data. Replace with actual content.'
-    },
-    more: {
-      title: 'More Data',
-      content: 'This is some static more data. Replace with actual content.'
-    }
-  };
+
+// Static data structure for different tabs
+const staticData = {
+  overview: {
+    title: 'Overview Data',
+    content: 'This is some static overview data. Replace with actual content.'
+  },
+  executive: {
+    title: 'Executive View Data',
+    content: 'This is some static executive view data. Replace with actual content.'
+  },
+  risk: {
+    title: 'Risk & Vulnerability View Data',
+    content: 'This is some static risk & vulnerability view data. Replace with actual content.'
+  },
+  more: {
+    title: 'More Data',
+    content: 'This is some static more data. Replace with actual content.'
+  }
+};
+
 const DashboardTabs = () => {
   const [activeTab, setActiveTab] = useState('overview');
-  const [data, setData] = useState(null);
+  const [data, setData] = useState(staticData.overview);
 
-  // Load static data based on active tab
+  // Update data based on the active tab
   useEffect(() => {
-    if (activeTab !== 'overview') {
-      setData(staticData[activeTab]);
-    }
+    setData(staticData[activeTab] || staticData.overview);
   }, [activeTab]);
 
   const handleTabClick = (tab) => {
@@ -36,108 +37,44 @@ const DashboardTabs = () => {
   };
 
   return (
-    <div>
+    <div style={{ width: '100%' }}>
       <div className="az-dashboard-nav">
         <nav className="nav">
-  <a
-    className={`nav-link ${activeTab === 'overview' ? 'active' : ''}`}
-    href="#overview"
-    onClick={() => handleTabClick('overview')}
-  >
-    Overview
-  </a>
-  <a
-    className={`nav-link ${activeTab === 'executive' ? 'active' : ''}`}
-    href="#executive"
-    onClick={() => handleTabClick('executive')}
-  >
-    Executive View
-  </a>
-  <a
-    className={`nav-link ${activeTab === 'risk' ? 'active' : ''}`}
-    href="#risk"
-    onClick={() => handleTabClick('risk')}
-  >
-    Risk & Vulnerability View
-  </a>
-  <a
-    className={`nav-link ${activeTab === 'more' ? 'active' : ''}`}
-    href="#more"
-    onClick={() => handleTabClick('more')}
-  >
-    More
-  </a>
-</nav>
-              <nav className="nav">
-                <a className="nav-link" href="/dashboard"><i className="far fa-save"></i> Save Report</a>
-                <a className="nav-link" href="/dashboard"><i className="far fa-file-pdf"></i> Export to PDF</a>
-                <a className="nav-link" href="/dashboard"><i className="far fa-envelope"></i>Send to Email</a>
-                <a className="nav-link" href="/dashboard"><i className="fas fa-ellipsis-h"></i></a>
-              </nav>
+          {Object.keys(staticData).map((tab) => (
+            <a
+              key={tab}
+              className={`nav-link ${activeTab === tab ? 'active' : ''}`}
+              href={`#${tab}`}
+              onClick={(e) => {
+                e.preventDefault();
+                handleTabClick(tab);
+              }}
+            >
+              {staticData[tab].title}
+            </a>
+          ))}
+        </nav>
+        <nav className="nav">
+          <a className="nav-link" href="/dashboard"><i className="far fa-save"></i> Save Report</a>
+          <a className="nav-link" href="/dashboard"><i className="far fa-file-pdf"></i> Export to PDF</a>
+          <a className="nav-link" href="/dashboard"><i className="far fa-envelope"></i> Send to Email</a>
+          <a className="nav-link" href="/dashboard"><i className="fas fa-ellipsis-h"></i></a>
+        </nav>
       </div>
       <div className="tab-content">
         {activeTab === 'overview' && <TabContents />}
-        {activeTab !== 'overview' && (
+        {activeTab === 'executive' && <ExecutiveView />}
+        {activeTab === 'risk' && <RiskView />}
+        {activeTab === 'more' && <TabContents />}
+        {Object.keys(staticData).indexOf(activeTab) === -1 && (
           <div>
-            {data ? (
-              <div>
-                <h2>{data.title}</h2>
-                <p>{data.content}</p>
-              </div>
-            ) : (
-              'Loading data...'
-            )}
+            <h2>{data.title}</h2>
+            <p>{data.content}</p>
           </div>
         )}
-      </div>
-      <div className="tab-content">
-        {activeTab === 'executive' && <ExecutiveView />}
-        {/* {activeTab !== 'executive' && (
-          <div>
-            {data ? (
-              <div>
-                <h2>{data.title}</h2>
-                <p>{data.content}</p>
-              </div>
-            ) : (
-              'Loading data...'
-            )}
-          </div>
-        )} */}
-      </div>
-      <div className="tab-content">
-        {activeTab === 'risk' && <RiskView />}
-        {/* {activeTab !== 'risk' && (
-          <div>
-            {data ? (
-              <div>
-                <h2>{data.title}</h2>
-                <p>{data.content}</p>
-              </div>
-            ) : (
-              'Loading data...'
-            )}
-          </div>
-        )} */}
-      </div>
-      <div className="tab-content">
-        {activeTab === 'more' && <TabContents />}
-        {/* {activeTab !== 'more' && (
-          <div>
-            {data ? (
-              <div>
-                <h2>{data.title}</h2>
-                <p>{data.content}</p>
-              </div>
-            ) : (
-              'Loading data...'
-            )}
-          </div>
-        )} */}
       </div>
     </div>
   );
 };
-
 
 export default DashboardTabs;
